feature(reintegrate-gestaltv7):  gestalt-v7 compability (package)

+++ src/main/java/org/terasology/machines/ui/HorizontalProgressBar.java
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.terasology.machines.ui;
 
-import org.terasology.math.JomlUtil;
-import org.terasology.utilities.Assets;
-import org.terasology.math.TeraMath;
 import org.joml.Vector2i;
-import org.terasology.rendering.assets.texture.TextureRegion;
+import org.terasology.engine.math.JomlUtil;
+import org.terasology.engine.rendering.assets.texture.TextureRegion;
+import org.terasology.engine.utilities.Assets;
+import org.terasology.math.TeraMath;
 import org.terasology.nui.Canvas;
 import org.terasology.nui.CoreWidget;
 import org.terasology.nui.LayoutConfig;
@@ -22,11 +22,12 @@ public class HorizontalProgressBar extends CoreWidget {
     @Override
     public void onDraw(Canvas canvas) {
         if (fillTexture != null) {
-            float result = (float) TeraMath.clamp(getValue());
+            float result = TeraMath.clamp(getValue());
 
             Vector2i size = canvas.size();
             int drawWidth = Math.round(result * fillTexture.getWidth());
-            canvas.drawTextureRaw(fillTexture, JomlUtil.rectangleiFromMinAndSize(0, 0, drawWidth, size.y), ScaleMode.STRETCH,
+            canvas.drawTextureRaw(fillTexture, JomlUtil.rectangleiFromMinAndSize(0, 0, drawWidth, size.y),
+                    ScaleMode.STRETCH,
                     0f, 0f, result, 1f);
         }
     }
