+++ src/main/java/org/terasology/machines/ui/VerticalProgressBar.java
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.terasology.machines.ui;
 
-import org.terasology.math.JomlUtil;
-import org.terasology.utilities.Assets;
-import org.terasology.math.TeraMath;
 import org.joml.Vector2i;
-import org.terasology.rendering.assets.texture.TextureRegion;
+import org.terasology.engine.math.JomlUtil;
+import org.terasology.engine.rendering.assets.texture.TextureRegion;
+import org.terasology.engine.utilities.Assets;
+import org.terasology.math.TeraMath;
 import org.terasology.nui.Canvas;
 import org.terasology.nui.CoreWidget;
 import org.terasology.nui.LayoutConfig;
@@ -22,12 +22,13 @@ public class VerticalProgressBar extends CoreWidget {
     @Override
     public void onDraw(Canvas canvas) {
         if (fillTexture != null && getValue() != null) {
-            float result = (float) TeraMath.clamp(getValue());
+            float result = TeraMath.clamp(getValue());
 
             Vector2i size = canvas.size();
             int drawHeight = Math.round(result * fillTexture.getHeight());
             int offsetHeight = Math.round((1 - result) * fillTexture.getHeight());
-            canvas.drawTextureRaw(fillTexture, JomlUtil.rectangleiFromMinAndSize(0, offsetHeight, size.x, drawHeight), ScaleMode.STRETCH,
+            canvas.drawTextureRaw(fillTexture, JomlUtil.rectangleiFromMinAndSize(0, offsetHeight, size.x, drawHeight)
+                    , ScaleMode.STRETCH,
                     0f, 0f, 1f, result);
         }
     }
